backend/requirements.txt
# Procura Backend Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
supabase>=2.0.0
postgrest>=0.16.0

# Task Queue (optional - for production)
celery>=5.3.0
redis>=5.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0
email-validator>=2.0.0

# Security
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# LLM Providers
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.0
tenacity>=8.2.0
structlog>=24.1.0

# Rate Limiting
slowapi>=0.1.9

# Testing (dev)
pytest>=8.0.0
pytest-asyncio>=0.23.0
